feat : 알림 생성 save 메소드 생성
Fixes: #16
